requirements.txt
cmake
Cython
munkres
numpy
opencv-python
packaging
pillow
pygame
pyparsing
pyyaml
